face_spoof_detection/requirements.txt
# requirements.txt
torch==2.0.1
torchvision==0.15.2
flask==2.3.2
Pillow==9.5.0
scikit-learn==1.2.2
matplotlib==3.7.1
tqdm==4.65.0
numpy==1.24.3